# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FactoryMind API
  version: 1.0.0
  summary: Industrial IoT monitoring API for Italian manufacturing SMEs.
  description: |
    FactoryMind exposes REST endpoints for real-time production monitoring:
    facilities, lines, devices (machines), telemetry metrics, alerts, OEE
    (Overall Equipment Effectiveness), and user administration.

    Most endpoints require a Bearer access token. Error responses follow
    RFC 7807 Problem Details (`application/problem+json`). State-changing
    endpoints are rate-limited (60 req/min/IP default).
  contact:
    name: FactoryMind Engineering
    email: support@factorymind.example
    url: https://factorymind.example/support
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:3002
    description: Local development (docker-compose)
  - url: https://api.factorymind.example
    description: Production
...
paths:
...
  /api/facilities:
    get:
      tags: [facilities]
      summary: List facilities
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: ok } }
    post:
      tags: [facilities]
      summary: Create facility
      security: [{ bearerAuth: [] }]
      responses: { '201': { description: created } }
...
  /api/devices:
    get:
      tags: [devices]
      summary: List devices (machines)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: facility_id
          schema: { type: string }
        - in: query
          name: line_id
          schema: { type: string }
      responses: { '200': { description: ok } }
    post:
      tags: [devices]
      summary: Register a device
      security: [{ bearerAuth: [] }]
      responses: { '201': { description: created } }
  /api/devices/{id}:
    get:
      tags: [devices]
      summary: Get a device
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: ok }
        '404': { $ref: '#/components/responses/Problem' }
  /api/devices/{id}/attestazione/piano4:
    post:
      tags: [devices]
      summary: Generate a Piano Transizione 4.0 attestazione document for the device
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Document generated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  eligibility: { type: boolean }
                  report: { type: object }
        '404': { $ref: '#/components/responses/Problem' }
...
  /api/alerts:
    get:
      tags: [alerts]
      summary: List active alerts
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: ok } }
  /api/alerts/{id}/acknowledge:
    post:
      tags: [alerts]
      summary: Acknowledge an alert
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: ok }
        '404': { $ref: '#/components/responses/Problem' }
  /api/alerts/{id}/resolve:
    post:
      tags: [alerts]
      summary: Resolve an alert
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: ok }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Problem:
      description: RFC 7807 error response.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
...